pyproject.toml
[tool.poetry]
name = "testing-the-untestable"
version = "0.1.0"
description = ""
authors = ["Michael Lerner <michael.lerner@ionix.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
ipytest = "^0.13.3"
pytest-sugar = "^0.9.7"
pytest-mock = "^3.12.0"
freezegun = "^1.2.2"
requests-mock = "^1.11.0"
requests = "^2.31.0"
pytest-httpx = "^0.27.0"
httpx = "^0.25.1"
mongomock-motor = "^0.0.21"
motor = "^3.3.2"
pydantic = "^2.5.1"
fastapi = "^0.104.1"
anyio = "3.7.1"
flask = "^3.0.0"
pytest-asyncio = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
